feat: Separate SchemanticType to a new package
Closes #222

The `schemantic` package has a heavy set of dependencies including
`analyzer` which makes it a poor candidate to be a shared dependency
across all implementations of JSON schema libaries that users may want
to use with `package:genkti`. Pull out `SchemanticType` to a new package
with fewer dependencies that may allow for other JSON schema libraries
to take a dependency and implement `SchemanticType` for compatibility.

Move `validate` to an extension method so that the implementation and
error interfaces can remain in `package:schemantic`.
