+++ packages/schemantic/pubspec.yaml
@@ -13,6 +13,7 @@ dependencies:
   code_builder: ^4.10.0
   dart_style: ^3.1.4
   json_schema_builder: ^0.1.3
+  schemantic_api: ^0.1.0
   source_gen: ^4.2.0
 
 dev_dependencies:
