+++ packages/schemantic_api/CHANGELOG.md
@@ -1,0 +1,4 @@
+## 0.1.0
+
+- Initial release. Expose the `SchemanticType` class with a slightly pruned API
+  and a significantly pruned dependency list.
