+++ packages/schemantic_api/README.md
@@ -1,0 +1,5 @@
+[![Pub](https://img.shields.io/pub/v/schemantic_api.svg)](https://pub.dev/packages/schemantic_api)
+
+A general-purpose Dart library for communicating JSON schemas. Most users will
+want to use `package:schemantic` instead. This package allows compatibility
+between multiple implementations of a JSON schema library.
