+++ packages/schemantic_api/test/custom_schema_test.dart
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'package:schemantic/schemantic.dart';
+import 'package:schemantic_api/schemantic.dart';
 import 'package:test/test.dart';
 
 void main() {
