+++ pubspec.yaml
@@ -19,6 +19,7 @@ workspace:
   - packages/genkit_shelf
   - packages/genkit_chrome
   - packages/schemantic
+  - packages/schemantic_api
   - packages/genkit_middleware
   - testapps/agentic_patterns
   - testapps/ai_features
